requirements/common.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
Django==5.1.4
django-allauth==0.61.1
django-appconf==1.0.6
django-compressor==4.5.1
django-environ==0.11.2
idna==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
rcssmin==1.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.2
sqlparse==0.4.4
stripe==9.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
django-tailwind-cli==2.13.0
django-htmx==1.18.0
django-render-block==0.9.2
django-extensions==3.2.3
django-simple-history==3.7.0
PyYaml==6.0.2
whitenoise==6.8.2
